requirements.txt
streamlit
tflite-runtime
numpy
Pillow
opencv-python
